fix: mark magic link as used instead of removing
The MIN_SECONDS_BETWEEN variable is used to protect against multiple
authentications in a short period of time.

However it is only able to protect against multiple authentication
failures. Successful authentications are not limited. A malicious actor
may perform a DoS or drive costs if she has access to a valid user
credential.

This fix enforces MIN_SECONDS_BETWEEN in both cases by creating an
`used` flag for each magic link. When created `used` is set to false.
To be used a magic link must have `used` as false. On use the `used`
is set to true. This will allow the MIN_SECONDS_BETWEEN to deny too
short creation even on successful uses.